--- src/main/java/com/umc/mada/calendar/CalendarController.java
@@ -1,4 +1,0 @@
-package com.umc.mada.calendar;
-
-public class CalendarController {
-}
