--- src/main/java/com/umc/mada/calendar/CalendarRepository.java
@@ -1,4 +1,0 @@
-package com.umc.mada.calendar;
-
-public interface CalendarRepository {
-}
