+++ src/main/java/com/umc/mada/calendar/controller/CalendarController.java
@@ -1,0 +1,4 @@
+package com.umc.mada.calendar.controller;
+
+public class CalendarController {
+}
