+++ src/main/java/com/umc/mada/calendar/domain/Calendar.java
@@ -1,0 +1,4 @@
+package com.umc.mada.calendar.domain;
+
+public class Calendar {
+}
