+++ src/main/java/com/umc/mada/calendar/repository/CalendarRepository.java
@@ -1,0 +1,4 @@
+package com.umc.mada.calendar.repository;
+
+public interface CalendarRepository {
+}
