+++ src/main/java/com/umc/mada/calendar/service/CategoryService.java
@@ -1,0 +1,4 @@
+package com.umc.mada.calendar.service;
+
+public class CategoryService {
+}
