--- src/main/java/com/umc/mada/category/CategoryRepository.java
@@ -1,4 +1,0 @@
-package com.umc.mada.category;
-
-public interface CategoryRepository {
-}
