+++ src/main/java/com/umc/mada/category/controller/CategoryController.java
@@ -1,0 +1,4 @@
+package com.umc.mada.category.controller;
+
+public class CategoryController {
+}
