+++ src/main/java/com/umc/mada/category/domain/Category.java
@@ -1,0 +1,4 @@
+package com.umc.mada.category.domain;
+
+public class Category {
+}
