+++ src/main/java/com/umc/mada/category/repository/CategoryRepository.java
@@ -1,0 +1,4 @@
+package com.umc.mada.category.repository;
+
+public interface CategoryRepository {
+}
