+++ src/main/java/com/umc/mada/category/service/CategoryService.java
@@ -1,0 +1,4 @@
+package com.umc.mada.category.service;
+
+public class CategoryService {
+}
