--- src/main/java/com/umc/mada/custom/Custom.java
@@ -1,4 +1,0 @@
-package com.umc.mada.custom;
-
-public class Custom {
-}
