--- src/main/java/com/umc/mada/custom/CustomController.java
@@ -1,4 +1,0 @@
-package com.umc.mada.custom;
-
-public class CustomController {
-}
