--- src/main/java/com/umc/mada/custom/CustomService.java
@@ -1,4 +1,0 @@
-package com.umc.mada.custom;
-
-public class CustomService {
-}
