+++ src/main/java/com/umc/mada/custom/controller/CustomController.java
@@ -1,0 +1,4 @@
+package com.umc.mada.custom.controller;
+
+public class CustomController {
+}
