+++ src/main/java/com/umc/mada/custom/domain/Custom.java
@@ -1,0 +1,4 @@
+package com.umc.mada.custom.domain;
+
+public class Custom {
+}
