+++ src/main/java/com/umc/mada/custom/repository/CustomRepository.java
@@ -1,0 +1,4 @@
+package com.umc.mada.custom.repository;
+
+public interface CustomRepository {
+}
