+++ src/main/java/com/umc/mada/custom/service/CustomService.java
@@ -1,0 +1,4 @@
+package com.umc.mada.custom.service;
+
+public class CustomService {
+}
