--- src/main/java/com/umc/mada/notice/NoticeController.java
@@ -1,4 +1,0 @@
-package com.umc.mada.notice;
-
-public class NoticeController {
-}
