--- src/main/java/com/umc/mada/notice/NoticeRepository.java
@@ -1,4 +1,0 @@
-package com.umc.mada.notice;
-
-public interface NoticeRepository {
-}
