+++ src/main/java/com/umc/mada/notice/controller/NoticeController.java
@@ -1,0 +1,4 @@
+package com.umc.mada.notice.controller;
+
+public class NoticeController {
+}
