+++ src/main/java/com/umc/mada/notice/domain/Notice.java
@@ -1,0 +1,4 @@
+package com.umc.mada.notice.domain;
+
+public class Notice {
+}
