+++ src/main/java/com/umc/mada/notice/repository/NoticeRepository.java
@@ -1,0 +1,4 @@
+package com.umc.mada.notice.repository;
+
+public interface NoticeRepository {
+}
