+++ src/main/java/com/umc/mada/notice/service/NoticeService.java
@@ -1,0 +1,4 @@
+package com.umc.mada.notice.service;
+
+public class NoticeService {
+}
