--- src/main/java/com/umc/mada/todo/Todo.java
@@ -1,4 +1,0 @@
-package com.umc.mada.todo;
-
-public class Todo {
-}
