--- src/main/java/com/umc/mada/todo/TodoRepository.java
@@ -1,4 +1,0 @@
-package com.umc.mada.todo;
-
-public interface TodoRepository {
-}
