--- src/main/java/com/umc/mada/todo/TodoService.java
@@ -1,4 +1,0 @@
-package com.umc.mada.todo;
-
-public class TodoService {
-}
