+++ src/main/java/com/umc/mada/todo/domain/Todo.java
@@ -1,0 +1,4 @@
+package com.umc.mada.todo.domain;
+
+public class Todo {
+}
