+++ src/main/java/com/umc/mada/todo/repository/TodoRepository.java
@@ -1,0 +1,4 @@
+package com.umc.mada.todo.repository;
+
+public interface TodoRepository {
+}
