+++ src/main/java/com/umc/mada/todo/service/TodoService.java
@@ -1,0 +1,4 @@
+package com.umc.mada.todo.service;
+
+public class TodoService {
+}
