--- src/main/java/com/umc/mada/user/User.java
@@ -1,4 +1,0 @@
-package com.umc.mada.user;
-
-public class User {
-}
