--- src/main/java/com/umc/mada/user/UserService.java
@@ -1,4 +1,0 @@
-package com.umc.mada.user;
-
-public class UserService {
-}
