+++ src/main/java/com/umc/mada/user/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.umc.mada.user.controller;
+
+public class UserController {
+}
