+++ src/main/java/com/umc/mada/user/domain/User.java
@@ -1,0 +1,4 @@
+package com.umc.mada.user.domain;
+
+public class User {
+}
