+++ src/main/java/com/umc/mada/user/repository/UserRepository.java
@@ -1,0 +1,4 @@
+package com.umc.mada.user.repository;
+
+public interface UserRepository {
+}
