+++ src/main/java/com/umc/mada/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.umc.mada.user.service;
+
+public class UserService {
+}
